Move palace plugin types to palace plugin
